requirements.txt
Flask
Pillow
PyPDF2
python-docx
Werkzeug